requirements.txt
aniso8601==9.0.1
attrs==21.4.0
click==8.1.0
Flask==2.1.0
flask-restx==0.5.1
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
jsonschema==4.4.0
MarkupSafe==2.1.1
pyrsistent==0.18.1
pytz==2022.1
six==1.16.0
Werkzeug==2.1.0
zipp==3.7.0
